Do not use LIMIT clause in UPDATE query
On database setups using replication, this can cause an 'Unsafe statement written to the binary log using statement format since BINLOG_FORMAT = STATEMENT. The statement is unsafe because it uses a LIMIT clause.' warning.

Since we are using an ID clause, we don't need to have a LIMIT clause because the ID is supposed to be unique.